Switch to pytest
Use pytest as framework for testing.
Sort tests into different files for better overview.
