+++ tox.ini
@@ -1,7 +1,10 @@
 [tox]
-envlist = py27, py36
+envlist = py{38,37,36,27}
 
 [testenv]
-deps = pytz>=2013a
+deps =
+    pytz>=2013a
+    pytest
+    pytest-mock
 whitelist_externals = make
-commands = make test
+commands = pytest
